First pass to enable GuidGenerator to generate multiple IIDs for a single .NET type.
This implementation is likely horribly non-performant because of its heavy usage of LINQ and iterator blocks, but it's a start.

Use arrays instead of a massive iterator block to implement GetSignatures.

Adapt LINQ-less implementation of GetAllPossibleSignatureCombinations from SO.

Adapt the recursive implementation instead of the non-recursive implementation from the same SO answer (easier to read).

Fix typo.

Enable multiple-signatures for PropertyChangedEventArgs.

Create a non-exception-throwing QI api to make supporting multiple IIDs for the same type more efficient.

Update codegen and CCW infrastructure to consume the multiple-PIIDs work.