Add `WindowsRuntimeReferenceAssemblyAttribute` (#2058)
* Add WindowsRuntimeReferenceAssemblyAttribute

Introduces a new attribute to identify assemblies as Windows Runtime reference assemblies. This attribute is intended for use by the CsWinRT generator and is analogous to ReferenceAssemblyAttribute, but specific to Windows Runtime projections.

* Clarify summary for WindowsRuntimeReferenceAssemblyAttribute

Updated the XML summary comment to better describe the attribute's purpose, clarifying that it applies to both reference and implementation assemblies containing generated Windows Runtime APIs from a .winmd file.