More tweaks to WinRT.Runtime and cswinrtgen (#2106)
* Refactor marshalling methods to use new marshallers

Updated ConvertToUnmanaged, ConvertToManaged, and Free methods to delegate to WindowsRuntimeUnknownMarshaller and WindowsRuntimeObjectMarshaller. This simplifies the code by removing manual null checks and logic, relying on the new marshalling APIs for object conversion and resource management.

* Refactor CCW unwrapping to use TryGetManagedObject

Replaces direct use of IsReferenceToManagedObject and ComInterfaceDispatch.GetInstance with WindowsRuntimeMarshal.TryGetManagedObject in object and unsealed object marshallers. This simplifies and unifies the logic for unwrapping managed objects from COM Callable Wrappers (CCWs).

* Use normal casts instead of 'Unsafe.As<T>(object)'

Needed for the type map to work properly

* Add casts in generated marshallers

* Remove unused System.Runtime.InteropServices imports

Eliminated unnecessary using directives for System.Runtime.InteropServices in WindowsRuntimeObjectMarshaller and WindowsRuntimeUnsealedObjectMarshaller to clean up the code.