Add `[WindowsRuntimeDefaultInterface]` type (#2108)
* Refactor Windows Runtime metadata name usage

Introduced WellKnownMetadataNames static class to centralize and cache well-known metadata names. Updated IsProjectedWindowsRuntimeType to use these constants for improved maintainability and consistency.

* Mark WindowsRuntimeMetadataAttribute as obsolete

Added Obsolete and EditorBrowsable attributes to WindowsRuntimeMetadataAttribute to discourage direct usage and hide it from editor intellisense. Also removed remarks indicating it is not meant for direct use.

* Add WindowsRuntimeDefaultInterfaceAttribute class

Introduces the WindowsRuntimeDefaultInterfaceAttribute to indicate the default interface for projected Windows Runtime class types. The attribute is marked obsolete and is intended for internal use.